There are now extra switches for what files to bind stdout and stderr to in traffic_manager and traffic_server.
Issues:
traffic_server keeps crashing, which I think it why traffic_manager is
crashing as well (b/c of a broken pipe read). However, traffic_manager
is not restarting traffic_server like it does on a clean build. I think
my changes may have something to do with it.

Todo:
Implement rotation for traffic.out, which shouldn't be hard given that
with the switches it should be fairly easy