Fix potential seg fault, remove some debug stuff
Narrowed down the cause for traffic_manager not restarting traffic_server
upon crash of traffic_server. The non-restart does not happen if you run
traffic_manager or traffic_server alone. Only if you run traffic_cop or
trafficserver. It would appear the problem is completely isolated to
traffic_cop. The new parameter passing is suspect, since it is the only
thing that was changed in traffic_cop as far as I can remember.

It looks like the initial start of traffic_manager by traffic_cop is fine,
just when it tries to get restarted everything goes to crap.