Make LogFile use BaseLogFile
Code currently compiles and runs, but access.log and error.log
is not appearing for some reason. Probably due to a bug somewhere.
Going to figure out this bug, then revert all the traffic.out
rotation changes.

traffic.out changes are going to be reverted because after a quick
chat with the team, it is not feasible to have that feature (at
least within the current architecture).