Add TM rotation logic, rotation config, and signal handlers in TS
Added rotation logic in TM to handle records.config configuration
as well as a function that will be used in the future to rotate
traffic.out.

Also added records.config configurations for rotating traffic.out.

Also added a signal handler in TS (SIGUSR2) for receiving
notification that traffic.out was rotated by TM.